+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 2.7.3 (December 16, 2020)
+
+ENHANCEMENTS:
+
+* Support logarithmic charts in dashboards with the "base" attribute within the chart mapping.
+
+## 2.7.2 (December 2, 2020)
+
+BUG FIXES:
+
+* Send sparkline_decimal_precision for dashboards even if 0
+
+ENHANCEMENTS:
+
+* Chart attributes can be set in both dashboard and dashboard json resources.
+
 ## 2.7.1 (October 26, 2020)
 
 BUG FIXES: